requirements.txt
# 安装当前包及其依赖
-e .

# 开发工具
pytest>=7.0.0
flake8>=6.0.0
black>=23.0.0

aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
boto3==1.37.3
botocore==1.37.3
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
courlan==1.3.2
dataclasses-json==0.6.7
dateparser==1.2.1
debugpy==1.8.12
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
frozenlist==1.5.0
h11==0.14.0
htmldate==1.9.3
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jusText==3.0.2
langchain==0.3.19
langchain-anthropic==0.3.8
langchain-community==0.3.18
langchain-core==0.3.40
langchain-deepseek==0.1.2
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph==0.3.0
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.11
lxml==5.3.1
lxml_html_clean==0.4.1
marshmallow==3.26.1
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
# Editable install with no version control (news_agent==0.1)
-e /Users/feiwang/Desktop/projects/news_agent
numpy==1.26.4
openai==1.64.0
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
tavily-python==0.5.1
tenacity==9.0.0
tiktoken==0.9.0
tld==0.13
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
zstandard==0.23.0
